added module informed iteration inferrence, transformed report into propertyconfig, added hooks for propertyconfig retrieval in extern (messy but that's life).  run() now returns the report, which is turned into a Properties python object

+++ src/core/engine/engine.h
@@ -100,7 +100,7 @@ class Engine {
         int environment_rate;
         bool learning_flag;
 
-        int iterations;
+        size_t iterations;
         bool verbose;
         Report *report;
